// src/app/shared/models/Dtos/facturaDetalleDto.ts
export interface FacturaDetalleDto{

        codigofacturadetalle : number;
        
        codigofactura: number;

        codigoproducto: string;

        actividadeconomica: string;

        codigoproductosin: number;

        descripcion: string;

        cantidad: number;

        unidadmedida: number;

        preciounitario: number;

        montodescuento: number;

        subtotal: number;

        cuenta: string;

        numeroserie: string;

        numeroimei: string;

        codigogrupo: string;
}